+++ package.json
@@ -22,7 +22,7 @@
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.0",
     "react-toastify": "^9.1.1",
-    "tailwindcss": "^3.2.6"
+    "tailwindcss": "^3.2.7"
   },
   "devDependencies": {
     "@types/react": "^18.0.26",
