+++ src/views/Home/TopProducts.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import "./top-products.css"
 import { useState,useEffect,useRef } from "react"
 
-import ProductCard from "./ProductCard"
-import ProductCard2 from './ProductCard2'
+import ProductCard from "./ProductCard2"
+import ProductCard2 from './ProductCard'
 
 
 import { TopProductType } from "../../types/authTypes"
@@ -31,19 +31,35 @@ const TopProducts = () => {
 
       {/* we have to show here top 4 products of 3 categories
       and that 3 categories will be decided by the 
-      number of order recived by the each category
+      number of orders recived by the each category
       i.e. top 3 category with highest sales */}
       
 
-      <p className="text-[28px] font-semibold blue-font">Our Top Products from Electronics</p>
+      <p className="text-[28px] font-semibold blue-font ">Top Products in Electronics</p>
       <div className="product-cards">
 
         {/* {topProduct?.map((item,index) => (
           <ProductCard productName={item.productName} productDescription={item.productDescription} key={`topProduct-${index}`}/>
         ))} */}
         {topProduct?.map((item,index) => (
-          <ProductCard2 productName={item.productName} productDescription={item.productDescription} productPrice={item.productPrice} productRating={item.avg_rating} key={`topProduct-${index}`}/>
+          <ProductCard2 productName={item.productName} productDescription={item.productDescription} productPrice={item.productPrice} productRating={item.avgRating} key={`topProduct-${index}`}/>
         ))}
+        
+      </div>
+      <p className="text-[28px] font-semibold blue-font">Top Products in Books</p>
+      <div className="product-cards">
+
+        {topProduct?.map((item,index) => (
+          <ProductCard2 productName={item.productName} productDescription={item.productDescription} productPrice={item.productPrice} productRating={item.avgRating} key={`topProduct-${index}`}/>
+        ))}
+        
+      </div>
+      <p className="text-[28px] font-semibold blue-font">Top Products in Clothes</p>
+      <div className="product-cards">
+        {topProduct?.map((item,index) => (
+          <ProductCard2 productName={item.productName} productDescription={item.productDescription} productPrice={item.productPrice} productRating={item.avgRating} key={`topProduct-${index}`}/>
+        ))}
+        
       </div>
     </div>
   )
