+++ src/views/Home/product-card.css
@@ -19,8 +19,23 @@
   cursor: pointer;
   width: 100%;
   font-size: 18px;
+  margin-top: 10px;
 }
 
 .card-button:hover {
   opacity: 0.7;
+}
+
+.like-button{
+  position: relative;
+    top: 10px;
+    left: 55%;
+}
+
+.badge {
+  background-color: red;
+  color: white;
+  padding: 4px 8px;
+  text-align: center;
+  border-radius: 5px;
 }
